Core: Allow passing top level dir modificationState
This avoids looking up the toplevel dir a second time inside the git
plugin. Since the modification state is used inside the project tree
this reduces the time to calculate the size hint in the project tree
from ~250ms to ~100ms for showing the tree for the first time on
windows.

Api: IVersionControl::modificationState(...)
 ->  IVersionControl::modificationState(..., const FilePath &topLevelDir = {})

Change-Id: Ice89a62265f35d5bd1f87c5c4f1b6c6971e2db8e
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: André Hartmann <aha_1980@gmx.de>