Utils: optimize FilePath::canonicalPath
The isDir check is not for free and requires around a third of the
runtime of the function. The result is used to check whether a flag
should be passed to CreateFile that has no effect on non Directories. So
just emit the check and always pass FILE_FLAG_BACKUP_SEMANTICS to
CreateFile.

Change-Id: I99e2f154a9218ebef979c071d0a6b4fdf059a5a4
Reviewed-by: Christian Stenger <christian.stenger@qt.io>