CMake: pass CMakeBuildTarget to CMakeTargetNode
Checking the flags for a project tree node results in checking
CMakeBuildSystem::supportsAction. Which potentially iterates over all
buildTargets for every CMakeTargetNode. Since the CMakeBuildTarget is
known at the point when the CMakeTargetNode is created we can pass it to
the node and directly access it from the node instead of iterating over
all BuildTargets.

Change-Id: Ibc6ef885c67a94023fe3a385a3aa32e3b0c22f42
Reviewed-by: Cristian Adam <cristian.adam@qt.io>