CppEditor: Speed up CppModelManager::getSignalSlotType()
This function is in the hot code path for completion and caused lag in
big files.
Skip function macro expansion, which is expensive and not needed here.
Speed-up on my machine is factor four in a release build and factor ten
in a debug build.
Arguably, it is still a bit too slow; will check for possible further
improvements.

Change-Id: I10ee22c074f5b76df161906c4211669646d82b6e
Reviewed-by: David Schulz <david.schulz@qt.io>