Python: fix crash when updating extra compilers in the pylsclient
Before 3d6eeea5dc40702cb11afffa72a1a05b7f0dfe31 we took the
extracompilers out of m_extraCompilers and readded the ones that are
still relevant. After the change we only append, which potentially
results in multiple compilers in that list, which in turn closes and
deletes some to early.

Change-Id: Ic3d199d84e49b54bcd15d2470c6276b6c500fd95
Reviewed-by: hjk <hjk@qt.io>