wizard: add checkbox for .qmlls.ini generation
Add a checkbox in the wizard to allow setting QT_QML_GENERATE_QMLLS_INI
in the CMakeLists.txt for new project. Disable the checkbox if 6.10 is
used as minimum version (6.10 and later does not need the .qmlls.ini
files).

In the CMakeLists.txt template, check that the minimum version is not
6.10 and whether the box was checked to set the
QT_QML_GENERATE_QMLLS_INI CMake variable. Note the double condition to
avoid setting the CMake variable in case the checkbox is checked and
disabled.

To avoid code duplication, save the condition whether the minimum qt
version is 6.8 into an option key as string and re-use it around by
expanding the string. (The options are not updated so saving the
condition directly in the string).

Also note that it seems that there are no version comparison method yet
available, doing '6.8 < 6.9' seemed to do float comparison, which fails
with qt version 6.10 because 6.10 ends up being smaller than 6.8...
Instead, check whether the minimum version is 6.8 (instead of checking
if < 6.10).

Fixes: QTCREATORBUG-33673
Change-Id: I5fc734d53819e4d8d63989153e0572ba34a19f14
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>