LanguageClient: fix outline for SymbolInformation
The lsp outline item constructed by a SymbolInformation did not setup
the client member, which was used to indicate whether an outline item is
valid. Fix this by defining a valid outline item by having a valid range
instead of a client set. After this the client member does not serve any
purpose so remove it as well.

Fixes: QTCREATORBUG-33668
Change-Id: Idd74fd72cbce74e71029751df5f935a51d99c134
Reviewed-by: Christian Stenger <christian.stenger@qt.io>