Android: Provide a simple device file access
This enables access to the file system of Android devices, for example in
the File System view and the Locator file system filter.

It uses the fallback access that uses 'find' and 'test' and so on,
wrapped with adb shell.

It adds a simple wrapper process interface around adb shell, to support
the base implementation for fileContents et al, which use plain Process.

The patch fixes issues on Android with find, which fails to substitute {}
for the found item if it isn't standing for itself (we directly
surrounded it with quotes, and that breaks the substitution).

It also fixes an issue with the 'ls' based fallback, which, like 'find',
returns a non-zero exit value if there are any kind of issues, like if
the user doesn't have permissions for some items.

Fixes: QTCREATORBUG-32697
Change-Id: I9369fad6f6acc55c4db524e8c8d4551137a7da42
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: hjk <hjk@qt.io>