ClangFormat: Fix potential use of wrong settings
Due to the way our editors are set up, the indenter gets created with an
empty file path. If client code retrieved the style during that time, an
effectively random value was returned, depending on Qt Creator's working
directory. Since the value is cached, we kept using it for the lifetime
of the indenter.

Change-Id: Ie1f515bd8af2126feb1da7f48a5dc0ec1e06f925
Reviewed-by: David Schulz <david.schulz@qt.io>