CMake: Improve UI text and documentation for staging installation
It was not clear from the UI what it actually does and what purpose it
serves. Rename the option to "Install into staging directory" and add a
informative tool tip. Also removes the label of the following path
chooser, since it is implicitly labeled by the preceeding check box.

Task-number: QTCREATORBUG-33640
Change-Id: Ieed73759e978a268757fe15b25cd06647153078e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>