Utils: Make DeviceFileAccess more thread safe
At the point where the DeviceFileAccess is aquired
(IDevice::fileAccess())
for usage we take a shared_ptr of it so that it can't be deleted from a
different thread while it is being used.

IDevicePrivate::fileAccess is changed to a SynchronizedValue to guard
against simultaneous read/write access in (set)fileAccess().

When using setFileAccessFactory the implementation is responsible to
ensure that calling the factory function is thread safe.

API: FilePath::fileAccess() now returns a shared_ptr<DeviceFileAccess>

Change-Id: I841a8b1270c4e174366984408a13710253d0f3c2
Reviewed-by: Eike Ziller <eike.ziller@qt.io>