Devices: Accept "device local paths" for tools in settings
The tools from the device tool aspects are usually supposed to be
located on the device.

The patch sets the "base directory" for the aspects to the device root,
so that paths without a scheme&host are resolved to be on the device,
which makes it much simpler to set them up by hand.

The patch changes the auto detection to also set paths without
scheme&host, which visually cleans this up further.

Change-Id: I5e0e497f478e2f6377ee0d97a12f92849e4c875b
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>