Core: Add support for freedesktop file managers
For FileUtils::showInGraphicalShell() for better support on Linux

Use 'org.freedesktop.FileManager1' [1] DBus interface to open user
default file manager and select file inside it's containing folder like
it's done on Windows with explorer.exe and Finder on macOS.

A file manager supporting this DBus service must be installed.
When service is not found or an error occurs, it falls back to old logic
which only shows the folder without file selection.

Tested with Dolphin and Gnome's Files.

Can be opted-out in System settings in case D-Bus creates issues like
timeouts.

[1]: https://freedesktop.org/wiki/Specifications/file-manager-interface/

Fixes: QTCREATORBUG-29368
Change-Id: I4e667ad3b27540659c927feb7e253a121028401d
Reviewed-by: hjk <hjk@qt.io>