CMake: Do not try to get capabilities for non-connected devices
During startup, when kits are restored, the project explorer asks all
kit aspects for "fix"ing themselves. The CMakeGenerator aspect asks for
the supported generators, which requests the capabilities by running the
CMake executable.

This can freeze Qt Creator for multiple seconds if there is a kit with a
remote build device with remote CMake that is currently not connected
but set to "auto-connect on startup".

Avoid running disconnected remote executables on startup in the main
thread with an explicit check if the device is connected.
Do not mess with the CMake generator in the kit, if it isn't.

Fixes: QTCREATORBUG-33736
Change-Id: I2291407434725ae2be0fc686dfccb0891a3dfd14
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Cristian Adam <cristian.adam@qt.io>