Locator/FileSystemFilter: Do not iterator twice
It was iterating once to get directories, another time to get files.
Just iterate once and sort into directories and files there.

As a side effect this avoids an issue with the fallback unix device file
access, which doesn't correctly sort symbolic links into directories and
files, leading to duplicate entries in Locator (once as a directory once
as a file).

Change-Id: I90986ec65571598cd44adf75ccb8a6a6198753a4
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>