Fix CMake tool on Docker devices
Amends 3fe33ba9a1e17ae3c0b2fc00c95d23068a82a35c

which made it important that build devices set their state to ReadyToUse
or at least Connected, because otherwise the capabilities for the CMake
tool are never fetched and the tool is considered "not suitable for Qt
Creator".

This cherry-picks the parts of d24e2c4304ddd5d42c9fa9705d34bf848afba9c9
from master branch that set the Docker device state and that prevent a
crash on shutdown when doing that.

Change-Id: I70661ce19b68f5042709137d422794938e141352
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>