RemoteLinux: LinuxDevicePrivate::setupShellPhase3
The FileAccess::deployAndInit function contain a lot of runBlocking
process runs, and since the code is executed on startup in the main
thread this increases the duration until Qt Creator becomes usable. For
me this can take up to 10 seconds. And since this is run per linux
remote device the startup time gets longer for each configured device

Change-Id: I6c051df4f38873d1ca2acc76974c14df7058ae3e
Reviewed-by: hjk <hjk@qt.io>