QtSupport: Correctly clone Qt versions with their detection type
Cloning does toMap-fromMap, and that lost the information about the
actual detection type (just supported "System" or "Manual").

Use DetectionSource to (de)serialize the detection source correctly.
Keep the old values for compatibility.

This fixes the that Qt versions with "Temporary" detection source were
saved to the settings and restored as "System" versions.

Change-Id: I76cf2f9fee25b86bafa652de1111f8634763304f
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>