QmlDesigner: Fix update of non-alias bound dynamic properties
There are some legitimate needs to call setupBackendValue() in
DynamicPropertyRow::handleDataChanged() even if call came via
DynamicPropertiesModel::updateItem(). For example, if a binding was
changed, setupBackendValue() call is needed to resolve the binding.

Instead of blocking setup for all properties, only block if for alias
properties, as those are the problematic ones.

Fixes: QDS-15993
Change-Id: I14569c492743b229e5bb38ce72ed68aa13067b3e
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>