CMake: Improve usability of the Qt Online Installer integration
This fixes the issue that it had a dependency on the availability of
maintenance tool during _initial_ configure time of the user project (by
setting a CMake variable for it).
It also automatically started and popped up the installer UI during
configure time, without even giving the user a chance to look at the
CMake error output.

This version still relies on Qt Creator's CMake helper scripts to give
us information about the missing Qt packages. But instead of starting
maintenance tool, it prints a warning to the CMake output, which then is
parsed by the CMake output parser.

  Qt packages are missing: 6.10.0, macos, qt.qt6.6100.addons.qtgraphs

  If you are using the Qt Online Installer, check the Issues view in Qt
  Creator for details.

The output parser adds a link "Install the missing component with the Qt
Online Installer" to the generated Task, that opens the installer UI
that is provided by the UpdateInfo plugin (via a URL handler), and is
available both in the Issues view and in the Task's tooltip e.g. in the
editor.

This is similar to how missing packages are handled by the Python
plugin. See PythonOutputLineParser::handleLine.

The patch also fixes an issue with Task::addLinkDetail, which didn't take
into account that other "detail" might already have been added, and
didn't support an alternative link text.

Change-Id: I9e4d20d498fea611682e71bbc4ead74108aa4354
Reviewed-by: Cristian Adam <cristian.adam@qt.io>
Reviewed-by: hjk <hjk@qt.io>