Reapply "RemoteLinux: LinuxDevicePrivate::setupShellPhase3"
This reverts commit 675b5d8dc6ccd04cd61efac8ddd45c701d13c489
and fixes a lock up at startup.

The reason for the previous lockup was that the linux device locked the
m_shellMutex before the whole setupShell machinery, but only unlocked it
after the new async deployAndInit was finished and reported via
onResultsReady to setupShellPhase3.

That meant that any file access before setupShellPhase3 finished would
wait for the lock in LinuxDevicePrivate::runInShell. But if that file
access was also blocking the main thread, the result of the async
deployAndInit was never handled on the main thread, since it is blocked.

The moment we set the fallback file access, we also need to unlock the
mutex that is needed for it to be used.

Change-Id: I4103f006d6e644464bd98ff151456724126b08ff
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>