SquishTests: Disable ClangFormat for rename macros test
After the renaming there might happen an internal additional
reformatting which is added to the undo stack.
So, undoing may undo the formatting first (no-op in this test
case), but the source itself would not get reverted.
Seems to depend additionally on the LLVM version in use
whether this additional formatting happens or not.
Avoid this by disabling ClangFormat plugin as the alternative
would be more costly.

Change-Id: I823a7b2a251ddf8025f49c751def3c0854906edd
Reviewed-by: Jukka Nokso <jukka.nokso@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>