Python: Unify the installation option from the issues view
Makes the style of the link for installing the missing package with pip
similar to the new CMake package one in
ac901e047bff7a86c709ddb78b6ac40056d87dc1
i.e. adapt the text, make the whole line a link, and use the Task API
for setting that up.

Change-Id: I1f7b82e0e95f633a3718ebdf44897671355ae7d6
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>