Find archiver program when variable CMAKE_AR is not set
In some projects with GCC or Clang as compiler, the program ar is used
but the variable CMAKE_AR is not set, so the usual mechanism to find
the archiver program does not work. I have added code to find the
archiver instead with find_program().

If there is another toolchain where CMAKE_AR is not automatically set,
the script outputs a warning and tries to continue. If this is the
case, we obviously need to add another special case to the script.

Fixes: QTCREATORBUG-33770
Change-Id: I1bb7abfc2b490d9b9199d6ba9838c389c2701298
Reviewed-by: David Schulz <david.schulz@qt.io>