Move build time to AppInfo and use QDateTime
So it can be used in other plugins (namely the licensechecker),
and it can be translated in the about dialog.

MSVC and GCC specify that __DATE__ will be "MMM dd yyyy" but with the
day left-filled with space instead of 0, and __TIME__ as "hh:mm:ss".
Though at least for GCC I do not see such a space in our previous builds
where we just displayed __DATE__ as is. Clang seems to do something
similar. In any case be safe and try formats without and with space.

Task-number: QTCREATORBUG-33751
Change-Id: If01b87728ba1dbb255ffff3f3b4368a4027b58c6
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>