CMake: Fix the path for run configuration templates
The RunConfigurationCreationInfo::projectFilePath is supposed to include
the file name of the project file, which is then shown in the "add run
config" dialog relative to the project root directory.

This also fixes a soft assert which actually checks exactly that. Make
that soft assert a bit more understandable by checking the variable
directly that it is supposed to check.

Change-Id: I72b910bc667cb4b30044106916fe98f3c37b3d2b
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Cristian Adam <cristian.adam@qt.io>