Utils: Introduce QtDesignSystemStyle
Simply overriding the paint method of QtcTabBar did not allow
implementing the design accurately. Also, dragging of tabes tab does not
work.

Proper styling of a tab element in a QTabBar needs to be done via
QStyle. This change introduces the QtDesignSystemStyle which initially
is only responsible for QtcTabBar. Eventually, more Qtc* widgets can be
handled by it.

Change-Id: Ib32081edcf4ddd427bfa429a4d4038427b9ea0ca
Reviewed-by: Cristian Adam <cristian.adam@qt.io>