Process/FilePath: Better default for process output encoding
Our Process implementation has a text encoding setting that is used for
APIs that request decoded process output.
The default delegates that decision to FilePath and therefore to the
FileAccess that is registered for that. But, for example during the
setup of the cmd bridge file access we want to run processes and retrieve
their output before we actually have a file access.

Previously that was handled by the UnavailableFileAccess, falling back
to Utf8, which we removed as unneeded in
e4b65ee6a84b04ada9f3d2c3af86a4f0a9d48cc2 . Make sure that the fallback
when we do not have a file access at all is also Utf8.

Change-Id: I1e50b774dfe184ba665aa2714feac03aade7e4b4
Reviewed-by: hjk <hjk@qt.io>