Add MCU output panel
Added a new panel to show the output of the qmlprojectexporter.
Does simple Ansi to HTML conversion to display the text in
the colors specified by the exporter. The issues panel now replaces
meaningless output with a link, when clicked opens the output panel
to make finding the actual error easier. This panel is only visible
in mcu projects. Note that the panel needs an icon for the status line.

Change-Id: I06d20cd846785eecfd5edac464ac20018fb677d8
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>