QmlPreview: Avoid duplicate error messages
If the preview service asks for a file called "qmldir", we don't want to
send a separate "error" message for each and every file called "qmldir"
in the project tree. We either have the exact match, then we should send
that. Or we don't, then we should send one error message.

Change-Id: Ibba00ec8f8bbedb11436b2c49c5155a8016878ab
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>