ProjectExplorer: Improve the search for QML files
Improve the performance of searching for QML files:
Running mimeTypeForFile is expensive on a lot of file, because for all
files all globs of all registered MIME types are checked.

Instead, for each file check it against a list of "QML extensions", which
is much faster (e.g. for Qt with declarative, tools, tests & examples
the old loop took ~900ms here, the new one ~80ms).

This is NOT equivalent to actually checking the MIME type: Even though
we construct the extensions from the MIME types, that only supports
simple globs, so manipulation of MIME types in the settings is limited
to that, and other MIME types could actually override the glob with a
higher priority so "*.qml" would no longer be classified as the qml MIME
type. But it is a reasonable optimization.

Change-Id: I8a19791692986791c06ffbdb2c9d8e78348dd71d
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>