Utils: Save memory by using QString COW in FilePath::fromString
An example is in IndexItem::create where a filename is provided
as a QString which is stored in a StringTable. Without this change that
StringTable is useless as the FilePath will always create a copy of the
QString data.

With this change the FilePath will use the QString data directly as long
as no modification is needed for the FilePath.

Change-Id: I9114e00649cc4e3bc454bdbdee8be7ae542419f8
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>