Git: Blame: Add option to remove file path and author
Git does not provide a way to omit the file path (only an option
to force it). For files that have been moved the file path is
included in the annotation so we need to filter it out manually.

For more flexibility, replace the code with regular expressions.
This is a bit slower, but makes the code more maintainable.

For example, the filtering of gitclient.cpp (4070 lines) takes
15 ms on my machine and the worst I have seen so far was 20 ms.

Fixes: QTCREATORBUG-33636
Change-Id: Ib82e02f1bd0d27248d807c374a61286ad96be93b
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>