iOS: Fix finding app bundle with single-platform Qt builds
In multi-platform builds, CMake emits the "${EFFECTIVE_PLATFORM_NAME}"
in the target artifact path, but in other cases it might not.
But in these cases the artifact path is actually the correct one and we
shouldn't try to construct the path ourselves with hardcoded platform
paths that then are wrong.

Change-Id: I0a48ce200a573cfbb70aa3b81dd3a5ab76be04d7
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>