CMakePM: Do not build `all` for a staging single target
When staging is used Qt Creator will add the "install" command to the
CMake building command line.

The "install" command has a dependency on the "all" target which would
mean that the whole project will be built instead of the just one target
that we selected.

This patchset makes sure that for Ninja and Makefile generators the "sub/
dir/target/all" and "sub/dir/target/install" targets are build.

Fixes: QTCREATORBUG-33580
Change-Id: Ia53ab48ac9e604d576992e3e386c76f5059aa150
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>