ProjectExplorer: Ensure key is present after Kit::setValue()
The QtProjectImporter test depends on that and is currently broken,
because debuggers are detected asynchronously and we don't have the
placeholder entries anymore.
I believe the optimization not to enter invalid values into the map was
accidental or at least not consistent, as we did not remove an existing
key when setValue() was called with an invalid QVariant.

Change-Id: I90c6c3843fbc52c6d157f1a0beaae7b65766e729
Reviewed-by: hjk <hjk@qt.io>