Core: Add settings apply/cancel dirty state management
This enables/disables Apply and Cancel buttons depending on
the dirty state of the settings pages.

For Aspect based pages, this is directly related to the aspect's
dirty/volatileValueChanged mechanism, i.e. correct.

For pages with "manually" created widgets (via setWidgetCreator)
pages are set to dirty if some well-known types of subwidgets
trigger certain changed signals. In this case, a page stays dirty
until Apply or Cancel are used. This happens also when the
changed subwidgets is manually changed back to the original value,
i.e. this is currently "overcommitting"

Note some settings pages didn't have a working Cancel implementation,
in the sense that the directly re-set values in the UI. This was
previously harmless as the page widget themselves were destroyed
when the settings dialog was closed. On re-open the widgets would
be recreating with new (saved) value.

This is now implemented by deleting/restoring widgets depending
on a RecreateOnCancel flag, which is set by default if and only
if the widget is created via setWidgetCreator.

For cases where setWidgetCreator is used together with a working
cancel() implementation, authors can use
IOptionsPage::setRecreateOnCancel(false).

The IOptionPage interface is now non-virtual, it acts very similar
to other factories now.

Switching away from a dirty page is prevented within the settings mode.

Current open questions are whether it would make to switch to
the previous mode on Apply/Cancel, and whether to block mode
switching in the presence of dirty pages.

Api: Removed IOptionsPage::{widget,apply,cancel}.
Replace with setWidgetCreator, and
IOptionsPageWidget::{apply,cancel} when needed.

Api: Removed IOptionsPage::matches(),
Partial replacement via setFixedKeywords.

Change-Id: I5e01bc3eda8d23a5012eb426bbcdaff7fb94f6a4
Reviewed-by: Eike Ziller <eike.ziller@qt.io>