Kits: Fix warning and persistence of "None" as a debugger
The "None" entry for the debugger in a kit references the default
DebuggerItem that has a default-constructed variant as its ID.
The ID of the DebuggerItem is saved in the profiles.xml for the kit.

The settings writing and reading did not support invalid variants as
values though. It would write a value with type name "", and reading
that back would write a warning and not restore the value. So a kit with
"None" set for the debugger appeared as having no debugger set and got
"fixed" with some debugger for the device. If available. If no debugger
was configured for the device, the warning would reappear again.

Add support for default-constructed variants to the settings saving,
which fixes both the warning and the resetting of the debugger even if
explicitly set to "None".

Change-Id: Ie446fa2153ab1d5dc39ea67fd41327bf8b443e68
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>