Git: Fix modification state for renamed files
In contrast to normal git file states like "M  foo",
the pattern for renamed files is "R  foo -> bar"
(two filenames separated by " -> "). This does not
match any file in the tree.

As the file system tree contains the new node,
we also have to insert the new name into the list.

Change-Id: I1b1038cea708954f490656690fd50dbe12de782b
Reviewed-by: Orgad Shaneh <orgads@gmail.com>